src/test/java/web/SpittleControllerTest.java
package web;

import org.junit.Test;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.ArrayList;
import java.util.Date;
import java.util.List;

import static org.junit.Assert.*;
import static org.junit.matchers.JUnitMatchers.hasItems;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;

/**
 * <p>文件名称：SpittleControllerTest </p>
 * <p>文件描述：</p>
 * <p>版权所有：版权所有(C)2011-2099 </p>
 * <p>公   司：口袋购物 </p>
 * <p>内容摘要：</p>
 * <p>其他说明：</p>
 * <p>完成日期：2017/2/18 </p>
 *
 * @author wangqiming
 */
public class SpittleControllerTest {

    @Test
    public void shouldShowRecentSpittles() throws Exception {
        List<Spittle> spittleList = createSpittleList(20);
        SpittleRepository mockRepository = mock(SpittleRepository.class);
        when(mockRepository.findSpittles(Long.MAX_VALUE,20)).thenReturn(spittleList);

        SpittleController spittleController = new SpittleController();
        MockMvc mockMvc = MockMvcBuilders.standaloneSetup(spittleController).build();
        mockMvc.perform(get("/spittles")).andExpect(view().name("spittlespage")).andExpect(model().attributeExists("spittleList"))
                .andExpect(model().attribute("spittleList",hasItems(spittleList.toArray())));
    }


    private List<Spittle> createSpittleList(int count) {
        List<Spittle> spittles = new ArrayList<Spittle>();
        for (int i = 0; i < count; ++i) {
            spittles.add(new Spittle("Spittle" + i, new Date()));
        }
        return spittles;
    }
}